requirements.txt
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.5
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.5
django-silk==4.1.0
djangorestframework==3.12.2
environs==9.3.2
flake8==3.9.0
Flask==1.1.2
gprof2dot==2021.2.21
idna==2.10
iniconfig==1.1.1
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.0
mccabe==0.6.1
numpy==1.20.1
packaging==20.9
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prettytable==2.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.0
Pygments==2.7.2
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
